move board + target into dirs
