+++ packages/svelte-flexiboards/src/lib/components/flexi-delete.svelte
@@ -6,7 +6,7 @@
 		FlexiDeleteController,
 		type FlexiDeleteClasses
 	} from '$lib/system/manage.svelte.js';
-	import { assistiveTextStyle, generateUniqueId } from '$lib/system/utils.svelte.js';
+	import { assistiveTextStyle, generateUniqueId } from '$lib/system/shared/utils.svelte.js';
 
 	/** @deprecated FlexiDelete's children props are now redundant and will be removed in v0.4. */
 	type FlexiDeleteChildrenProps = {
