+++ packages/svelte-flexiboards/src/lib/components/flexi-target-loader.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { getInternalFlexitargetCtx } from '$lib/system/target.svelte.js';
+	import { getInternalFlexitargetCtx } from '$lib/system/target/index.js';
 
 	// This is a simple component that triggers the initial layout load.
 	// We must do this after instantiating all other components, otherwise we won't have established
