+++ packages/svelte-flexiboards/src/lib/system/board/types.ts
@@ -1,0 +1,7 @@
+import type { FlexiTargetDefaults } from '../target/types.js';
+import type { FlexiWidgetDefaults } from '../widget/types.js';
+
+export type FlexiBoardConfiguration = {
+	widgetDefaults?: FlexiWidgetDefaults;
+	targetDefaults?: FlexiTargetDefaults;
+};
